arch_sigreturn: change tracee address type from long to unsigned long
* linux/arm/arch_sigreturn.c (arch_sigreturn): Change addr variable
type from long to unsigned long.
* linux/ia64/arch_sigreturn.c (arch_sigreturn): Likewise.
* linux/m68k/arch_sigreturn.c (arch_sigreturn): Likewise.
* linux/powerpc/arch_sigreturn.c (arch_sigreturn): Likewise.
* linux/s390/arch_sigreturn.c (arch_sigreturn): Likewise.
* linux/sparc/arch_sigreturn.c (arch_sigreturn): Likewise.
* linux/tile/arch_sigreturn.c (arch_sigreturn): Likewise.
* linux/crisv10/arch_sigreturn.c (arch_sigreturn): Change the type
of addr and regs variables from long to unsigned long.
* linux/mips/arch_sigreturn.c (arch_sigreturn): Change addr variable
type from long to kernel_ureg_t.
* linux/x86_64/arch_sigreturn.c (arch_sigreturn): Change the type
of offset and addr variables from long to kernel_ureg_t.

+++ linux/crisv10/arch_sigreturn.c
@@ -1,13 +1,14 @@
 static void
 arch_sigreturn(struct tcb *tcp)
 {
-	long regs[PT_MAX + 1];
+	unsigned long regs[PT_MAX + 1];
 
-	if (ptrace(PTRACE_GETREGS, tcp->pid, NULL, (long)regs) < 0) {
+	if (ptrace(PTRACE_GETREGS, tcp->pid, NULL, regs) < 0) {
 		perror_msg("sigreturn: PTRACE_GETREGS");
 		return;
 	}
-	const long addr = regs[PT_USP] + offsetof(struct sigcontext, oldmask);
+	const unsigned long addr =
+		regs[PT_USP] + offsetof(struct sigcontext, oldmask);
 
 	tprints("{mask=");
 	print_sigset_addr_len(tcp, addr, NSIG / 8);
